Resolve requested changes + modify input manager
For the input manager, I switched most of its methods to static for ease of use instead of always call the GetInstance. I also made a little modification to the Controller so that we can get the cursor back when pressing VK_ESCAPE. We should also implement a IsKeyReleased method.